pyproject.toml
[tool.poetry]
name = "starship-command"
version = "0.1.0"
description = "The REST API server of the game"
authors = ["Mark Meliksetyan <markmelix@gmail.com>", "Anton Grishin <anton.ingrish@gmail.com>"]
license = "GPL-3.0-or-later"
readme = "README.md"
packages = [{include = "starship"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.2.3"
flask-wtf = "^1.1.1"
flask-sqlalchemy = "^3.0.3"
flask-login = "^0.6.2"
pyruvate = "^1.2.1"
python-dotenv = "^1.0.0"
werkzeug = "^2.2.3"
psycopg2 = "^2.9.5"
pyyaml = "^6.0"

[build-system]
requires = ["poetry-core>=1.4.0"]
build-backend = "poetry.core.masonry.api"
